NOTE #1: Yesterday's merge was incomplete (SVN silently failed before 17391 due to file name case changes).

NOTE #2: Because a couple EHR SQL scripts were edited in this range of revisions, anyone who updated and ran after yesterday's merge may have an EHR schema that can't be upgraded.  We'll worry about this if it happens.

SVN r17529 |2011-10-07 00:03:53 +0000

+++ ehr/resources/schemas/dbscripts/postgresql/EHR_Lookups-11.10-11.101.sql
@@ -2703,8 +2703,8 @@ test varchar(255) ,
 species varchar(255) ,
 gender varchar(255) ,
 age_class int2 DEFAULT NULL,
-ref_range_min numeric(255),
-ref_range_max numeric(255),
+ref_range_min double precision,
+ref_range_max double precision,
 RowId serial NOT NULL,
 
 CONSTRAINT PK_lab_test_range PRIMARY KEY (rowid)
